main/requirements.txt
Django==2.2.4
Pillow==6.1.0
djangorestframework==3.10.2